End of Lesson 2 code
These are the intended results from the end of Lesson 2 of Udacity's C++ Nanodegree program. In truth, much of it assumes you already know C++ coding (they never once mention the && operator on line 19 in their content) and can not only remember something mentioned once several virtual 'pages' ago, but can identify how it is applied (if written for you) or needs to be applied (if it's part of the intended solution of a problem) in your final draft for this-or-that function.